Move implementation detail off public API
