+++ packages/chat-stateful-client/src/ChatClientState.ts
@@ -109,24 +109,3 @@ export type ChatObjectMethodNames<TName extends string, T> = {
 // eslint complains on all uses of `Function`. Using it as a type constraint is legitimate.
 // eslint-disable-next-line @typescript-eslint/ban-types
 export type ChatMethodName<T, K extends keyof T> = T[K] extends Function ? (K extends string ? K : never) : never;
-
-/**
- * Method to decide at runtime if a string is an error target.
- */
-export const isChatErrorTarget = (target: string): boolean => {
-  for (const targetPrefix in ChatErrorTargetPrefixes) {
-    const [prefix, obj] = targetPrefix;
-    if (target.startsWith(prefix)) {
-      target = target.substring(prefix.length);
-      if (obj[target] !== undefined && typeof obj[target] === 'function') {
-        return true;
-      }
-    }
-  }
-  return false;
-};
-
-const ChatErrorTargetPrefixes = [
-  ['ChatClient.', ChatClient],
-  ['ChatThreadClient.', ChatThreadClient]
-];
